fix metadata and seo for game pages
